+++ tests/test_data_selector.py
@@ -1,11 +1,13 @@
 """Unit tests for DataSelector."""
 
-from warnings import catch_warnings
+from typing import List
 
+from pydantic import BaseModel, Field
 from trame.app import get_server
 from trame_server.core import Server
 
-from nova.trame.view.components.ornl import NeutronDataSelector
+from nova.mvvm.trame_binding import TrameBinding
+from nova.trame.view.components import DataSelector
 from nova.trame.view.theme import ThemedApp
 
 
@@ -19,30 +21,48 @@ def test_data_selector() -> None:
         def create_ui(self) -> None:
             with super().create_ui() as layout:
                 with layout.content:
-                    input = NeutronDataSelector(v_model="test", extensions=[".tiff"])
+                    input = DataSelector(v_model="test", directory="/", extensions=[".tiff"])
                     assert input.v_model == "test"
+                    assert input._directory == "/"
                     assert input._extensions == [".tiff"]
-                    assert input._model.state.facility == ""
-                    assert input._model.state.instrument == ""
-                    assert input._model.state.experiment == ""
 
-                    input.set_state(facility="HFIR", instrument="CG-2", experiment="IPTS-27744")
-                    assert input._model.state.facility == "HFIR"
-                    assert input._model.state.instrument == "CG-2"
-                    assert input._model.state.experiment == "IPTS-27744"
+    MyTrameApp()
 
-                    with catch_warnings(record=True) as captured_warnings:
-                        input.set_state(facility="NSS")
-                        assert str(captured_warnings[0].message).startswith("Facility 'NSS' could not be found.")
 
-                    with catch_warnings(record=True) as captured_warnings:
-                        NeutronDataSelector(v_model="test", facility="HIFR")
-                        assert str(captured_warnings[0].message).startswith("Facility 'HIFR' could not be found.")
+def test_parameter_bindings() -> None:
+    class TestModel(BaseModel):
+        v_model: List[str] = Field(default=[])
+        directory: str = Field(default="")
+        extensions: List[str] = Field(default=[])
+        prefix: str = Field(default="")
+        refresh_rate: int = Field(default=0)
+        select_strategy: str = Field(default="")
 
-                    with catch_warnings(record=True) as captured_warnings:
-                        NeutronDataSelector(v_model="test", facility="SNS", instrument="BL1B")
-                        assert str(captured_warnings[0].message).startswith(
-                            "Instrument 'BL1B' could not be found in 'SNS'."
-                        )
+    class MyTrameApp(ThemedApp):
+        def __init__(self, server: Server = None) -> None:
+            self.server = get_server(None, client_type="vue3")
+            super().__init__(server=self.server)
+
+            self.create_binding()
+            self.create_ui()
+
+        def create_binding(self) -> None:
+            self.test_obj = TestModel()
+
+            binding = TrameBinding(self.server.state)
+            self.test_binding = binding.new_bind(self.test_obj)
+            self.test_binding.connect("test_ds")
+
+        def create_ui(self) -> None:
+            with super().create_ui() as layout:
+                with layout.content:
+                    DataSelector(
+                        v_model=("test_ds.v_model", ["test.txt"]),
+                        directory=("test_ds.directory", "/"),
+                        extensions=("test_ds.extensions", [".txt"]),
+                        prefix=("test_ds.prefix", "bin"),
+                        refresh_rate=("test_ds.refresh_rate", 15),
+                        select_strategy=("test_ds.select_strategy", "page"),
+                    )
 
     MyTrameApp()
